% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datorama_process.R
\name{datorama_process}
\alias{datorama_process}
\title{Process Data Streams}
\usage{
datorama_process(access_token = NULL, dataStreamIds = NULL,
  startDate = NULL, endDate = NULL, create = FALSE)
}
\arguments{
\item{access_token}{Authorization string that is found within the Datorama platform}

\item{dataStreamIds}{A list of integers relationg to Data Stream IDs to process}

\item{startDate}{A string -> format YYYY-MM-DD}

\item{endDate}{A string -> format YYYY-MM-DD}

\item{create}{A boolean indicating whether to create a new stream or not}
}
\value{
A list containing the processing request's results
}
\description{
Process all the data streams listed in `dataStreamIds` between the dates
`startDate` and `endDate` (format YYYY-MM-DD). The optional field `create`
determines whether it's the first time the data streams are being processed.
Default value is false. Data streams being processed for the first time
(i.e. create: true) will also have a SmartLenses dashboard created for them.
This is designed to allow flexibility when creating complex setups via API.
It is the API user's responsibility to determine if they want processed data
streams to trigger SmartLenses creation.
}
\details{
Note: Used for API streams only.
}
\examples{
\dontrun{
datorama_process(acces_token = "aaaaaaaa-bbbb-cccc-dddd-eeeeeeeeeeee",
                 dataStreamIds = list(348937, 34289),
                 startDate = "2018-01-01",
                 endDate = "2018-01-20",
)
}
}
