% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datorama_query.R
\name{datorama_query}
\alias{datorama_query}
\title{A query constructor for the Datorama Query API}
\usage{
datorama_query(access_token = NULL, workspaceId = NULL, dateRange = NULL,
  startDate = NULL, endDate = NULL, measurements = NULL,
  dimensions = NULL, groupDimensionFilters = NULL,
  stringDimensionFilters = NULL, stringDimensionFiltersOperator = NULL,
  numberMeasurementFilter = NULL, sortBy = NULL, sortOrder = NULL,
  topResults = NULL, groupOthers = NULL, topPerDimension = NULL)
}
\arguments{
\item{access_token}{Authorization string that is found within the Datorama platform}

\item{workspaceId}{A string containing the workspaceId}

\item{dateRange}{A string containing date range for query}

\item{startDate}{A string -> format YYYY-MM-DD}

\item{endDate}{A string -> format YYYY-MM-DD}

\item{measurements}{A list of lists including attributes to include -> list(list(name = "Investment"))}

\item{dimensions}{A list of dimensions to be included -> list("Day", "Site Name")}

\item{groupDimensionFilters}{A list of lists including dimension, operator -> "IN", "NOT IN", and vals for filtering}

\item{stringDimensionFilters}{A list of lists including dimension, operator -> "EQUALS", "NOT_EQUALS", "NOT_EQUALS_CASE_SENSITIVE", "CONTAINS", "NOT_CONTAINS", "STARTS_WITH", "ENDS_WITH", "IS_EMPTY", "IS_NOT_EMPTY", and val for filtering}

\item{stringDimensionFiltersOperator}{A string for string dimension filtering -> "AND", "OR"}

\item{numberMeasurementFilter}{A list of lists including filedName (measurement name), operator -> "EQUALS", "NOT_EQUALS", "GREATER", "GREATER_EQUALS", "LESS", "LESS_EQUALS", "IS_NULL", "IS_NOT_NULL", "IS_NAN", "IS_NOT_NAN" and val for filtering}

\item{sortBy}{A string including name of dimension/measurement to sort by}

\item{sortOrder}{A string containing the order of the sort -> "ASC", "DESC"}

\item{topResults}{A string containing the number of results to return}

\item{groupOthers}{Boolean determining whether to group all results not in topResults}

\item{topPerDimension}{Boolean determining whether to return topResults per dimension}
}
\value{
A tibble with the query's response
}
\description{
Refer to Datorama's Developer portal for futher information. In
order to use this package your account will have to have access
to the Datorama Query API, which is a paid feature.
}
\examples{
\dontrun{
datorama_query(acces_token = "aaaaaaaa-bbbb-cccc-dddd-eeeeeeeeeeee",
               workspaceId = "271",
               dateRange = "CUSTOM",
               startDate = "2017-11-01",
               endDate = "2017-11-20",
               measurements = list(list(name = "Clicks"),
                              list(name = "Impressions")),
               dimensions = list("Day", "Site Name")
)
}
}
