% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datorama_auth.R
\name{datorama_token}
\alias{datorama_token}
\title{Authenticating yourself to Datorama}
\usage{
datorama_token(email, password)
}
\arguments{
\item{email}{A string containing the email associated with your Datorama account}

\item{password}{A string containing the password associated with your Datorama account}
}
\value{
A string with your our API token
}
\description{
Authenticating yourself to Datorama
}
\examples{
\dontrun{
datorama_token("someEmail@gmail.com", "password1234")
}
}
